ported more classes to eagleradapter

+++ src/main/java/net/lax1dude/eaglercraft/beta/FilesystemUtils.java
@@ -1,5 +1,6 @@
 package net.lax1dude.eaglercraft.beta;
 
+import java.util.Collection;
 import java.util.function.Consumer;
 
 import net.lax1dude.eaglercraft.EaglerAdapter;
@@ -8,10 +9,21 @@ import net.lax1dude.eaglercraft.adapter.EaglerAdapterImpl2.FileEntry;
 public class FilesystemUtils {
 	
 	public static void recursiveDeleteDirectory(String dir) {
-		EaglerAdapter.listFiles(dir, true, true).forEach(new Consumer<FileEntry>() {
+		Collection<FileEntry> lst = EaglerAdapter.listFiles(dir, true, true);
+		lst.forEach(new Consumer<FileEntry>() {
 			@Override
 			public void accept(FileEntry t) {
-				EaglerAdapter.deleteFile(t.path);
+				if(!t.isDirectory) {
+					EaglerAdapter.deleteFile(t.path);
+				}
+			}
+		});
+		lst.forEach(new Consumer<FileEntry>() {
+			@Override
+			public void accept(FileEntry t) {
+				if(t.isDirectory) {
+					EaglerAdapter.deleteFile(t.path);
+				}
 			}
 		});
 		EaglerAdapter.deleteFile(dir);
